added functionality and style for ev info page
